Extracted data into index.js
