+++ src/App.js
@@ -16,8 +16,8 @@ const App = (props) => {
                 <Navbar/>
                 <div className='app-wrapper-content'>
                     <Routes>
-                        <Route path='/dialogs' element={<Dialogs />} />
-                        <Route path='/profile' element={<Profile />} />
+                        <Route path='/dialogs' element={<Dialogs dialogsData={props.dialogsData} messages={props.messages}/>}/>
+                        <Route path='/profile' element={<Profile posts={props.posts}/>}/>
                     </Routes>
                 </div>
                 {/*<ProfileInfo/>*/}
