+++ src/components/Profile/MyPosts.jsx
@@ -2,16 +2,10 @@ import React from 'react'
 import styles from './MyPosts.module.css';
 import Post from "./MyPosts/Post/Post";
 
-const MyPosts = () => {
-
-    let posts = [
-        {id: 1, message: 'I am going to the sea', likesCount: "2"},
-        {id: 2, message: 'I\'ve bought the car', likesCount: "5"},
-        {id: 3, message: 'That\'s my first post', likesCount: "3"},
-    ]
+const MyPosts = (props) => {
 
     let postsElements =
-        posts.map(post => <Post message={post.message} likesCount={post.likesCount}/>)
+        props.posts.map(post => <Post message={post.message} likesCount={post.likesCount}/>)
 
     return <div className={styles.postsBlock}>
         <h3>My posts</h3>
